test: Seperate tests with partial class, covering regular, editor-only, and shared sections
